backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
Pillow
opencv-python
pydantic
python-multipart